last task

+++ client/src/app/members/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from 'src/app/_models/message';
 import { Pagination } from 'src/app/_models/pagination';
+import { ConfirmService } from 'src/app/_services/confirm.service';
 import { MessageService } from 'src/app/_services/message.service';
 
 @Component({
@@ -16,7 +17,7 @@ pageNumber = 1;
 pageSize = 5;
 loading = false;
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService,private confirmService:ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -32,10 +33,15 @@ loadMessages(){
 }
 
 deleteMessage(id:number){
-  this.messageService.deleteMessage(id).subscribe(()=>{
-    console.log('after http delete subscribe');
-    this.messages.splice(this.messages.findIndex(m=>m.id ===id),1);
+  this.confirmService.confirm('Confirm delete message','This cannot be undone').subscribe(result=>{
+    if(result)
+    {
+      this.messageService.deleteMessage(id).subscribe(()=>{
+      this.messages.splice(this.messages.findIndex(m=>m.id ===id),1);
   })
+    }
+  })
+  
 }
 
 pageChanged(event:any)
